+++ src/pious/pio/util.py
@@ -1,53 +1,5 @@
 from collections import namedtuple
 
-TEST_TREE = r"C:\PioSOLVER\Reports\SimpleTree\SRP\b25\BTNvBB"
-
-
-ranks = {
-    "2": 2,
-    "3": 3,
-    "4": 4,
-    "5": 5,
-    "6": 6,
-    "7": 7,
-    "8": 8,
-    "9": 9,
-    "T": 10,
-    "J": 11,
-    "Q": 12,
-    "K": 13,
-    "A": 14,
-}
-
-
-ranks_rev = {
-    2: "2",
-    3: "3",
-    4: "4",
-    5: "5",
-    6: "6",
-    7: "7",
-    8: "8",
-    9: "9",
-    10: "T",
-    11: "J",
-    12: "Q",
-    13: "K",
-    14: "A",
-}
-
-
-def ahml(rank):
-    if rank == 14:
-        return "A"
-    if rank > 9:
-        return "H"
-    if rank > 5:
-        return "M"
-    else:
-        return "L"
-
-
 def color_texture(texture):
     """
     Return a coloration of a texture
@@ -97,10 +49,6 @@ def marker_size_from_high_card(flop, max_size=None, min_size=10):
     size = min_size * factor ** (r - 2)
     return size
 
-
-def card_tuple(c):
-    r, s = c.strip()
-    return ranks[r], s
 
 
 Info = namedtuple("Info", ["player", "node_id", "line", "starting_stacks"])
